pyproject.toml
[project]
name = "z-laconic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm",
    "python-dotenv"
]
